Fixed issue with zig-zagging
Fixed issue with zig-zagging due to incorrect heuristic calculation making the weights for zig-zagging incorrect being 2 instead of root(2)

+++ A Star Pathfinding/A Star Pathfinding/Node.cs
@@ -10,9 +10,9 @@ namespace A_Star_Pathfinding
         public int y;
         public Node parentNode;
         public List<Node> neighbours;
-        public int fCost;
-        public int gScore;
-        public int hScore;
+        public double fCost;
+        public double gScore;
+        public double hScore;
         public bool isBarrier;
 
         public Node(int x, int y)
